supprimer autowired supplémentaires pour les tests
