+++ src/main/java/com/example/managershop/ManagerShopApplication.java
@@ -25,15 +25,8 @@ import java.util.stream.Stream;
 public class ManagerShopApplication //implements CommandLineRunner
 {
 
-	//@Autowired
-	private CategorieService categorieService;
-	//@Autowired
-	private RoleService roleService;
 
-	public ManagerShopApplication(CategorieService categorieService, RoleService roleService) {
-		this.categorieService = categorieService;
-		this.roleService = roleService;
-	}
+
 
 
 
@@ -52,29 +45,7 @@ public class ManagerShopApplication //implements CommandLineRunner
 		});
 	}*/
 
-	@Bean
-	CommandLineRunner start(/*AccountService accountService, RoleService roleService, CategorieService categorieService*/){
-		return  args -> {
 
-			/*Stream.of("Role1","Role2","Role3","Role4","Role5").forEach(r->{
-				    accountService.save(Role.builder().nameRole(r).levelPriorite((int)(Math.random() * 6)).build());
-*/
-				Stream.of("Categorie1","Categorie2","Categorie3","Categorie4","Categorie5","Categorie6").forEach(c->{
-					try {
-						categorieService.addCategory(Categorie.builder().nomCat(c).archived(false).build());
-					} catch (NullException e) {
-						e.printStackTrace();
-					}
-				});
-
-				categorieService.findAll().forEach(cat->{
-					System.out.println(cat.getNomCat());
-
-					//System.out.println(mapEntityToDto.CategorieToCategorieDTO(new Categorie("test")).toString());
-			});
-
-		};
-	}
 
 
 	@Bean
